+++ react/typings/vtex.pixel-manager/PixelContext.d.ts
@@ -1,3 +1,0 @@
-declare module 'vtex.pixel-manager/PixelContext' {
-  export const usePixel
-}